server/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
cryptography==41.0.2
Deprecated==1.2.14
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.100.0
frozenlist==1.4.0
future==0.18.3
gcloud==0.18.3
googleapis-common-protos==1.59.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httptools==0.6.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.5.0
MarkupSafe==2.1.3
multidict==6.0.4
oauth2client==4.1.3
orjson==3.9.2
passlib==1.7.4
Pillow==10.0.0
protobuf==4.24.0rc2
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.0.2
pydantic-extra-types==2.0.0
pydantic-settings==2.0.1
pydantic_core==2.1.2
pyparsing==3.1.0
Pyrebase4==4.7.1
python-dotenv==1.0.0
python-jose==3.3.0
python-jwt==4.0.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
tinydb==4.8.0
tuspy==1.0.1
typing_extensions==4.7.1
ujson==5.8.0
urllib3==1.26.16
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
yarl==1.9.2